+++ wrapper.pyx
@@ -25,9 +25,9 @@ from cython.view cimport array as cvarray
 import numpy as np
 #we define the header for the C function we want to call, it has to match the C function signature
 cdef extern from "min_sum.h":
-    void min_sum(float *L, const int *pcm_matrix, 
+    void min_sum(double *L, const int *pcm_matrix, 
                             int* syndrome, int size_checks, int size_vnode, 
-                            float *Lj, float alpha, int num_it, int * error_computed)
+                            double *Lj, double alpha, int num_it, int * error_computed)
 
 
 #in order to use the wrapper is necessary to flatten the matrices first
@@ -38,10 +38,10 @@ def compute_min_sum_wrapper(L,non_zero,syndrome,size_checks,size_vnodes,priors,a
         non_zero = np.ascontiguousarray(non_zero)
     
     #create memoryviews so we can pass numpy array pointers to C function
-    cdef float[::1] L_array = L
+    cdef double[::1] L_array = L
     cdef int[::1] non_zero_array = non_zero
     cdef int[::1] syndrome_array = syndrome
-    cdef float[::1] priors_array = priors
+    cdef double[::1] priors_array = priors
     cdef int[::1] error_computed_array = error_computed
